feat(install): Enabled proxy tunnelling on HTTP/HTTPS requests
By simply using 'caw' module to create the HTTP agent with proxy settings
we are able to use 'pnpm' behind a basic corporate proxy (could not test
it extensively on any kind of proxy).
I had to disable 'sepia' caching in case a proxy configuration is found
because it was not behaving fine with the new HTTP agent setup.