feat: add support for version range in github resolver
This allows to depend on version ranges from github repos, for example:

    rstacruz/pnpm#>=1.0.0

This makes pnpm capable to install versioned packages from github. The first
step to liberate from npm registry!

First we try to treat github ref as version range if it is validated as such.
Then we get a list of all tags in the repo and try tp match all of them (which
are valid semver versions) against the range.

Otherwise we just fallback to the prev behaviour and treat ref as ref in github
repo — either commit, branch or tag.

Note that we need to add the same code to git resolver, probably using `git
ls-remote` command to get a list of tags for a repo.