fix: ignore cycles of length 1 in dependency graph when sorting
Cycles of length 1 in the dependency graph ie., packages that depend on
themselves, confuse the 'graph-sequencer' function that is used to order the
packages topologically.

Sometimes, it makes sense to make packages depend on themselves eg., to install
their exposed scripts in `node_modules/.bin`, so we should either fix the
`graph-sequencer` or apply this patch.

An example for this behavior is:

- Package 'a' depends on 'b' and 'c'
- Package 'b' depends on 'b'
- Package 'c' depends on 'b' and 'c'

Now run:

	pnpm recursive --filter=a... [some command]

**Expected behavior**:

- Packages are visited in order [['b'], ['c'], ['a']], where no parallel
  execution is possible.

**Actual behavior**:

- Packages are visited in order [['b'], ['a'], ['c']].

This patch fixes that behavior.