fix: ignore filtered packages when sorting topologically
This is another issue with the graph-sequencer producing strange results
when
sorting.

Case that is fixed here:

- Package 'a' depends on 'b' and 'c'
- Package 'd' depends on 'a'
- Package 'e' depends on 'a', 'b' and 'c'

Command that has the wront behavior:

	pnpm recursive --filter=...a [command]

**Expected behavior**:

Packages are executed in order ['a', ['d', 'e']] ('d' and 'e' can be
executed
in parallel.)

**Actual behavior**:

Packages are executed in order [['e', 'a', 'd']]

**Fix**:

Remove the unused dependencies 'b' and 'c' from the graph when sorting.