Catch exception from package.json containing an invalid name
This is probably not where you really want to catch this exception - this is just to say that normalize-package-json does throw exceptions if, for example, package.json contains a name with spaces in it instead of hyphens. I don't know where you do want to catch this exception, but the stack trace (from loud-rejection?) makes it look like something has gone wrong inside of pnpm instead of simply unexpected data.