feat(outdated): add long option, no details by default
Previously, by default the `outdated` commands displayed a “Details” column which usually contained the URL to the package repo. As the URL could be quite long, on narrow terminals it would cause the table to wrap in a way which ruined the table borders.

This commit hides the “Details” column by default, and displays the column if `--long` is passed to the command.