feat(outdated): add --no-table option
This PR adds a `--no-table` option that prints out
the info on each package vertically concatenated with no borders
and a blank line between each package.

Note that the `--long` option may also be used in conjunction
with `--no-table` to add the details at the end of each set
of package info.