feat(pnpm): don't fail 'recursive run' if no selected packages have script
This changes the behavior of of `pnpm recursive run ...` so that it no
longer throws an error when run against a subset of packages and none of
the selected packages have the provided script.

Previously, running a command like `pnpm m run --filter foo build` would
throw an error if the `foo` package did not have a `build` script, even
if other workspace packages have a `build` script.

With this change, running `pnpm recursive run --filter foo build` will
not throw an error.

Note that if the provided package selectors match all packages (e.g.
`'*'`), then `pnpm recursive run` will fail if none of the packages have
the script. This is because the behavior depends on whether we're
running against a subset of packages or not, _not_ whether any package
selectors were provided.