docs(README): add mature section
pnpm is still unknown for many developers. As a result, when they
first find it, their first assumption is that it is "a new package
manager".

So we should mention in the README that pnpm has years of development
invested into it, and it is used by some big companies.