feat: no symlink
This is a continuation of #1874.
Setting `symlink=false` in an `.npmrc` file will alow to experiment
with Plug'n'Play, which for now is only possible by manually
generating the `.pnp.js` file using the `@pnpm/lockfile-to-pnp` package.

This is probably a temporary setting. In the future, there will be
something like `enable-pnp` instead.