fix(lockfile): change sorting of keys in lockfile
Make it more deterministic and prevent unnecessary churn in the lockfile.
It is useful when looking for affected projects after changes to the lockfile.
When the changes happen among various developers using various locale,
the locale specific sorting causes changes in the lockfile for extraneous projects.
Therefore, it could cause running unnecessary CI jobs, builds, releases and other
work that would not be needed otherwise.