chore(meta-updater): remove none determinism in lint:meta flow
The meta-updater package performs the diff checks on
all the required files **simultaneously**

- https://github.com/pnpm/meta-updater/blob/
  61179a083206acbd90c24a7ce7d337cb838a46f9/src/index.ts#L107

This could cause the **lazy** calculation of the mock ports to be incorrect.
- https://github.com/pnpm/pnpm/blob/
  ebf6ee09bd7295d4d4cbfec1876d7f4bff09bf76/.meta-updater/src/index.ts#L151

Thus causing random and inconsistent local build errors.

This commits solves the issue by calculating in **advance** the expected
port numbers, thus removing the assumption that package.json files
would be diffed checked in alphabetical order.