ci: introduce a centralized GHA check/gate job (#5723)
This patch is intended to lower the maintenance burden of having to
manually go through each matrix subjob name in the branch protection
repository settings. It allows to only include the check job in the
branch protection and it will robustly determine if the dependencies
have succeeded or not.

It is currently mostly serves the Python ecosystem in projects like
aiohttp, cryptography, open edX, pip etc. But I've also seen other
communities picking it up lately, like the AWS Rust SDK and even the
engine powering https://dev.to, to my surprise. Strictly speaking, it
is agnostic.

Ref: https://github.com/marketplace/actions/alls-green#why