fix(dependencies-hierarchy): initialize resultHeight to null
In a debugger, I noticed a test package with no dependencies return a
height of `0` from getTree. By the comment in `DependencyInfo`, this
should instead be `null` since the dependencies array is empty.

It turns out this is the case when `deps == null`. The `getTree`
function early returns with a `null` height:

```ts
if (deps == null) {
  return { dependencies: [], isPartiallyVisited: false, height: null }
}
```

However, when `deps` is `{}`, `resultHeight` gets initialized to `0`.
The subsequent `.forEach` loop then iterates over the non-empty object.