chore(pd): make esbuild find `js-yaml` correctly
For some reason, esbuild uses the upstream version of `js-yaml` rather
than the @zkochan fork. This throws a little plugin into esbuild's
resolution to resolve `js-yaml` using Node's resolvers instead, which
seems to find the correct version.

There may be a better way to do this, especially considering esbuild
does fine in the normal build of pnpm as far as I can tell... though
that may be because `tsc` does the first pass.