fix(get-context): avoid excess purge prompts
pnpm sometimes needs to purge modules folders (node_modules) when
certain configuration changes (in other words, when it needs to start
from scratch).

This prompt, problematically, was getting triggered asynchronously for
each importer in the workspace that needed a purge, which would
sometimes lead to gargantuan prompts instead of one consolidated prompt.

The fix is to queue up which importers need a purge, then fire the
prompter all at once.

Closes #7320