fix: process relative transitive peer references
The change is to recalculate local workspace references based on the
position of their lockfile from the project root.
This is important in case of transitive local peer dependencies lying
in a different levels in directory structure.
ERR_PNPM_LINKED_PKG_DIR_NOT_FOUND happened during deploy command
execution for such a projects.

Fixes: #8159, #5042, #4514