fix(reviewing/license-scanner): update license-scanner to support hoisted node modules
Applications built using electron and electron forge require node
modules to be hoisted to the root of the project.

Libraries that depend on the output of `pnpm licenses` have issues
resolving packages as the paths returned assume the node-linker is set
to isolated (or undefined).

This change refactors the existing behaviour that generates the path of
a given package. The abstracted method retrieves the config from the
project root to determine if a different path should be returned when
node-linker=hoisted.

This change may not cover all options for node-linker, however it does
imporve on the existing behaviour. It shouldn't break anything more than
what it already would be.

Closes: GitHub Issue #8589