fix: changes local-resolver to support absolute paths (#9761)
* feat: changes local-resolver to support absolute paths

Previously absolute paths were being turned to relative paths, but if the file:
specifier is 'file:/path/to/file', and the users are using a shared network
storage, this relative path requires that the users all use the same
local folder structure. Instead, using an absolute path as the specifier
allows them to have the source code anywhere, and the absolute path will
be resolved consistently.

Enabled via the `preserveAbsolutePaths` option.

* chore: changeset

* feat: add preserve absolute paths option

* docs: add changesets

* fix: also update the 'dependencyPath', add test for that case

---------

Co-authored-by: Zoltan Kochan <z@kochan.io>